requirements.txt
anthropic>=0.40.0
mem0ai>=0.1.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
